wip: more relocations
